fix: fire update

+++ simulator/simulator-software/src/operation/Operation.java
@@ -57,13 +57,9 @@ public class Operation {
     }
 
     public void notifyOnSite() {
-        if (this.status == OperationStatus.ON_ROAD) {
-            this.counter++;
-            System.out.println("COUNTER: " + this.counter);
-            if (this.counter < 30) return;
+        if (this.status == OperationStatus.ON_ROAD && LocalDateTime.now().isAfter(this.start.plusMinutes(3))) {
             this.status = OperationStatus.ON_SITE;
             OperationRepository.notifyOnSite(this.id);
-            this.counter = 0;
         }
     }
 
@@ -76,10 +72,7 @@ public class Operation {
     }
 
     public void notifyOnFinished() {
-        if (this.status == OperationStatus.RETURNING) {
-            this.counter++;
-            System.out.println("COUNTER: " + this.counter);
-            if (this.counter < 30) return;
+        if (this.status == OperationStatus.RETURNING && LocalDateTime.now().isAfter(this.start.plusMinutes(5))) {
             OperationRepository.notifyFinished(this.id);
             FireRepository.remove(this.id);
         }
